rename storage.py to sql_storage.py and move it to common/storage
