+++ google/cloud/security/iam/server.py
@@ -20,13 +20,13 @@ import time
 from concurrent import futures
 import grpc
 
+from google.cloud.security.common.storage.sql_storage import Storage
 from google.cloud.security.iam.client import ClientComposition
 from google.cloud.security.iam import db
 from google.cloud.security.iam.dao import ModelManager, create_engine
 from google.cloud.security.iam.explain.service import GrpcExplainerFactory
 from google.cloud.security.iam.playground.service import GrpcPlaygrounderFactory
 from google.cloud.security.iam.inventory.service import GrpcInventoryFactory
-from google.cloud.security.iam.inventory.storage import Storage
 
 
 # TODO: The next editor must remove this disable and correct issues.
