+++ tests/iam/api_tests/inventory_env_test.py
@@ -18,6 +18,7 @@ import unittest
 import time
 from sqlalchemy import event
 
+from google.cloud.security.common.storage.sql_storage import Storage
 from google.cloud.security.common.util.threadpool import ThreadPool
 from google.cloud.security.iam.explain.service import GrpcExplainerFactory
 from google.cloud.security.iam.inventory.service import GrpcInventoryFactory
@@ -25,7 +26,6 @@ from google.cloud.security.iam.playground.service import GrpcPlaygrounderFactory
 from google.cloud.security.iam.dao import ModelManager
 from google.cloud.security.iam.client import ClientComposition
 from google.cloud.security.iam import db
-from google.cloud.security.iam.inventory.storage import Storage
 
 from tests.iam.api_tests.api_tester import ApiTestRunner
 from tests.iam.utils.db import create_test_engine, cleanup
